refactor(mobile contacts): better semantic name for entrypoint
