+++ client-app/src/apps/contactMobile.ts
@@ -2,12 +2,12 @@ import '../Bootstrap';
 
 import {XH} from '@xh/hoist/core';
 import {AppContainer} from '@xh/hoist/mobile/appcontainer';
-import {AppComponent} from '../examples/mobileContacts/AppComponent';
-import {AppModel} from '../examples/mobileContacts/AppModel';
+import {AppComponent} from '../examples/contactMobile/AppComponent';
+import {AppModel} from '../examples/contactMobile/AppModel';
 import {AuthModel} from '../core/AuthModel';
 
 XH.renderApp({
-    clientAppCode: 'mobileContacts',
+    clientAppCode: 'contactMobile',
     clientAppName: 'XH Contact Mobile',
     componentClass: AppComponent,
     modelClass: AppModel,
