+++ client-app/src/examples/contactMobile/AppComponent.ts
