+++ client-app/src/examples/contactMobile/AppModel.ts
@@ -1,7 +1,7 @@
 import {XH} from '@xh/hoist/core';
 
 // No change in contact service. Pull from the desktop version to avoid redundancy
-import {ContactService} from '../../examples/contact/svc/ContactService';
+import {ContactService} from '../contact/svc/ContactService';
 import {BaseAppModel} from '../../BaseAppModel';
 
 export class AppModel extends BaseAppModel {
