+++ client-app/src/examples/contactMobile/ContactsPageModel.ts
