+++ client-app/src/examples/contactMobile/cmp/FavoriteButton.ts
