+++ client-app/src/examples/contactMobile/details/ContactDetails.ts
