+++ client-app/src/examples/contactMobile/details/ContactDetailsModel.ts
